[racingcar]test: divide test codes to match classes of main
